+++ sentry_streams_k8s/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.0.6
+
+### New Features ✨
+
+- (pipeline-step) Add runner log level support by @fpacifici in [#272](https://github.com/getsentry/streams/pull/272)
+
+## 0.0.5
+
+- Upgrade sentry-streams by @fpacifici in [#271](https://github.com/getsentry/streams/pull/271)
+
 ## 0.0.4
 
 - Add optional liveness probe and health volume to PipelineStep manifest by @fpacifici in [#234](https://github.com/getsentry/streams/pull/234)